NRFX-4892: Add missing casts in nrfx_timer driver

+++ drivers/src/nrfx_timer.c
@@ -289,8 +289,8 @@ void nrfx_timer_extended_compare(nrfx_timer_t const *   p_instance,
     NRFX_ASSERT(m_cb[p_instance->instance_id].state == NRFX_DRV_STATE_INITIALIZED);
 
     nrfy_timer_shorts_disable(p_instance->p_reg,
-        (NRF_TIMER_SHORT_COMPARE0_STOP_MASK  << cc_channel) |
-        (NRF_TIMER_SHORT_COMPARE0_CLEAR_MASK << cc_channel));
+        (uint32_t)(NRF_TIMER_SHORT_COMPARE0_STOP_MASK  << cc_channel) |
+        (uint32_t)(NRF_TIMER_SHORT_COMPARE0_CLEAR_MASK << cc_channel));
 
     nrfy_timer_shorts_enable(p_instance->p_reg, timer_short_mask);
